Solved Java host not always getting all last AD values bug
DESCRIPTION OF BUG AND SOLUTION:

The Rabbit had no problems retrieving the last AD values from the master
PIC but issues occurred when the Rabbit tried to transmit the value on to
the host; success did occur but was random.

The problem was caused by a mistake in handleGetAllLastADValuesPICCmd.
Variable i was never being assigned a value, but was still being passed to
sendPacketOfBuffersViaSocket() as the length of buf1. It worked sometimes
because there were times when the location in the stack reserved for
variable i already had a 0 in it (which was the proper length of buf1).

Since handleGetAllLastADValuesPICCmd() puts no values in buf1, the length
was changed to 1 instead of 12.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>