Xmt of Monitor Packet after Monitor Mode started works!!!!
Problem was in Master PIC. Refer to that repository for more details.

Added a temporary test variable so that the monitor byte received from the
Master PIC can be easily read in debug mode.

Also removed bail out in reSyncSP(). I had it commented out while fixing
the monitor packet issue and am not sure if it is necessary or if
uncommenting will negatively affect results.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>